refactor(tests): remove obsolete test modules and consolidate tests into dedicated files for Redis and PostgreSQL adapters, enhancing organization and maintainability
