+++ crates/adapter/src/lib.rs
@@ -41,51 +41,3 @@ pub const VERSION: &str = env!("CARGO_PKG_VERSION");
 
 /// Library name
 pub const NAME: &str = env!("CARGO_PKG_NAME");
-
-// Load environment variables from .env file for tests
-#[cfg(test)]
-#[ctor::ctor]
-fn init() {
-    dotenv::dotenv().ok();
-}
-
-#[cfg(test)]
-mod test_helpers {
-    use std::env;
-
-    /// Get Redis URL from environment variable with fallback to default
-    pub fn get_test_redis_url() -> String {
-        env::var("REDIS_URL")
-            .unwrap_or_else(|_| "redis://default:redispw@localhost:55000".to_string())
-    }
-
-    /// Get PostgreSQL URL from environment variable with fallback to default
-    pub fn get_test_postgres_url() -> String {
-        env::var("POSTGRES_URL").unwrap_or_else(|_| {
-            "postgresql://postgres:postgres@localhost:5432/dbx_test".to_string()
-        })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_version_exists() {
-        assert!(!VERSION.is_empty(), "Version should be defined");
-    }
-
-    #[test]
-    fn test_redis_url_from_env() {
-        use test_helpers::get_test_redis_url;
-
-        // Test that the function returns a valid URL
-        let url = get_test_redis_url();
-        assert!(!url.is_empty(), "Redis URL should not be empty");
-        assert!(
-            url.starts_with("redis://"),
-            "Redis URL should start with redis://"
-        );
-    }
-}
