+++ crates/adapter/src/tests/redis_tests/mod.rs
@@ -1,13 +1,10 @@
 //! Redis adapter test modules
 
-/// Redis factory tests
-mod factory_tests;
-
-/// Redis admin primitive tests
-mod admin_tests;
-
 /// Redis bitmap primitive tests
 mod bitmap_tests;
+
+/// Redis factory tests
+mod factory_tests;
 
 /// Redis set primitive tests
 mod set_tests;
