+++ crates/dbx-api/src/server.rs
@@ -635,360 +635,3 @@ pub enum ServerError {
     #[error("Server runtime error: {0}")]
     ServerRuntime(String),
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use axum::body::Body;
-    use axum::http::{Method, Request, StatusCode};
-    use std::sync::Arc;
-    use tower::util::ServiceExt;
-
-    /// Helper function to set up required environment variables for tests
-    fn setup_test_env() {
-        std::env::set_var(
-            "JWT_SECRET",
-            "test-jwt-secret-that-is-at-least-32-characters-long-for-security",
-        );
-        // Set up mock backend for testing
-        std::env::set_var("DBX_BACKEND_1_URL", "mock://localhost:6379");
-        std::env::set_var("DBX_BACKEND_1_PROVIDER", "mock");
-        std::env::set_var("DBX_BACKEND_1_NAME", "test_backend");
-        std::env::set_var("DBX_BACKEND_1_POOL_SIZE", "10");
-        std::env::set_var("DBX_DEFAULT_BACKEND", "test_backend");
-        std::env::set_var("HOST", "127.0.0.1");
-        std::env::set_var("PORT", "3000");
-    }
-
-    /// Helper function to clean up test environment variables
-    fn cleanup_test_env() {
-        std::env::remove_var("JWT_SECRET");
-        std::env::remove_var("DBX_BACKEND_1_URL");
-        std::env::remove_var("DBX_BACKEND_1_PROVIDER");
-        std::env::remove_var("DBX_BACKEND_1_NAME");
-        std::env::remove_var("DBX_BACKEND_1_POOL_SIZE");
-        std::env::remove_var("DBX_DEFAULT_BACKEND");
-        std::env::remove_var("HOST");
-        std::env::remove_var("PORT");
-        std::env::remove_var("CREATE_DEFAULT_ADMIN");
-        std::env::remove_var("DEFAULT_ADMIN_USERNAME");
-        std::env::remove_var("DEFAULT_ADMIN_PASSWORD");
-    }
-
-    /// Helper function to create AppState for tests with error handling
-    async fn create_test_app_state() -> AppState {
-        crate::test_helpers::create_test_app_state()
-            .await
-            .expect("Failed to create test AppState")
-    }
-
-    #[tokio::test]
-    async fn test_create_app_state_success() {
-        let _app_state = create_test_app_state().await;
-        // App state creation test - compiles and succeeds
-    }
-
-    #[tokio::test]
-    #[serial_test::serial]
-    async fn test_create_app_state_with_default_admin() {
-        setup_test_env();
-        std::env::set_var("CREATE_DEFAULT_ADMIN", "true");
-        std::env::set_var("DEFAULT_ADMIN_USERNAME", "admin");
-        std::env::set_var("DEFAULT_ADMIN_PASSWORD", "admin123");
-
-        let result = AppState::new(None).await;
-        // Default admin creation might fail in some test environments (concurrent tests, permissions, etc.)
-        // The important thing is that the application handles the configuration
-        match result {
-            Ok(_) => {
-                // Admin creation succeeded
-            }
-            Err(ServerError::UserStoreInitialization(_)) => {
-                // Configuration parsed correctly, but user creation failed
-            }
-            Err(ServerError::Configuration(_)) => {
-                panic!("Configuration should have been valid");
-            }
-            Err(_) => {
-                // Other errors are also acceptable in test environments
-            }
-        }
-
-        cleanup_test_env();
-    }
-
-    #[tokio::test]
-    #[serial_test::serial]
-    async fn test_create_app_state_missing_admin_credentials() {
-        setup_test_env();
-        std::env::set_var("CREATE_DEFAULT_ADMIN", "true");
-        std::env::set_var("DEFAULT_ADMIN_USERNAME", "admin");
-        std::env::remove_var("DEFAULT_ADMIN_PASSWORD");
-
-        let result = AppState::new(None).await;
-        assert!(result.is_err());
-
-        if let Err(ServerError::Configuration(_)) = result {
-            // Expected error type
-        } else {
-            panic!("Expected Configuration error");
-        }
-
-        cleanup_test_env();
-    }
-
-    #[tokio::test]
-    async fn test_create_app_with_cors() {
-        let app_state = create_test_app_state().await;
-        let app_config = create_test_app_config();
-        let app = create_app_with_config(app_state, app_config);
-
-        // Test CORS preflight on auth endpoint (should work because CORS is applied there)
-        let auth_request = Request::builder()
-            .method(Method::OPTIONS)
-            .uri("/auth/login")
-            .header("Origin", "http://localhost:3000")
-            .header("Access-Control-Request-Method", "POST")
-            .body(Body::empty())
-            .unwrap();
-
-        let auth_response = app.oneshot(auth_request).await.unwrap();
-        assert_eq!(auth_response.status(), StatusCode::OK);
-
-        // Test OPTIONS on health endpoint (should return 405 as it doesn't support OPTIONS)
-        let app_state = create_test_app_state().await;
-        let app = create_app(app_state).await.unwrap();
-
-        let health_request = Request::builder()
-            .method(Method::OPTIONS)
-            .uri("/health")
-            .body(Body::empty())
-            .unwrap();
-
-        let health_response = app.oneshot(health_request).await.unwrap();
-        assert_eq!(health_response.status(), StatusCode::OK);
-    }
-
-    #[tokio::test]
-    async fn test_health_check_endpoint() {
-        let app_state = create_test_app_state().await;
-        let app_config = create_test_app_config();
-        let app = create_app_with_config(app_state, app_config);
-
-        let request = Request::builder()
-            .method(Method::GET)
-            .uri("/health")
-            .body(Body::empty())
-            .unwrap();
-
-        let response = app.oneshot(request).await.unwrap();
-        assert_eq!(response.status(), StatusCode::OK);
-    }
-
-    #[tokio::test]
-    async fn test_middleware_chain() {
-        let app_state = create_test_app_state().await;
-        let app_config = create_test_app_config();
-        let app = create_app_with_config(app_state, app_config);
-
-        let request = Request::builder()
-            .method(Method::GET)
-            .uri("/health")
-            .body(Body::empty())
-            .unwrap();
-
-        let response = app.oneshot(request).await.unwrap();
-        assert_eq!(response.status(), StatusCode::OK);
-    }
-
-    #[tokio::test]
-    async fn test_cors_configuration() {
-        let app_state = create_test_app_state().await;
-        let app_config = create_test_app_config();
-        let app = create_app_with_config(app_state, app_config);
-
-        // Test CORS preflight on auth endpoint (should work - CORS enabled for browser access)
-        let auth_preflight_request = Request::builder()
-            .method(Method::OPTIONS)
-            .uri("/auth/login")
-            .header("Origin", "http://localhost:3000")
-            .header("Access-Control-Request-Method", "POST")
-            .body(Body::empty())
-            .unwrap();
-
-        let response = app.clone().oneshot(auth_preflight_request).await.unwrap();
-        assert_eq!(response.status(), StatusCode::OK);
-    }
-
-    #[tokio::test]
-    async fn test_not_found_endpoint() {
-        let app_state = create_test_app_state().await;
-        let app_config = create_test_app_config();
-        let app = create_app_with_config(app_state, app_config);
-
-        let request = Request::builder()
-            .method(Method::GET)
-            .uri("/nonexistent")
-            .body(Body::empty())
-            .unwrap();
-
-        let response = app.oneshot(request).await.unwrap();
-        assert_eq!(response.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn test_protected_route_without_auth() {
-        let app_state = create_test_app_state().await;
-        let app_config = create_test_app_config();
-        let app = create_app_with_config(app_state, app_config);
-
-        let request = Request::builder()
-            .method(Method::GET)
-            .uri("/api/v1/data/test-key")
-            .body(Body::empty())
-            .unwrap();
-
-        let response = app.oneshot(request).await.unwrap();
-        assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn test_admin_route_without_auth() {
-        let app_state = create_test_app_state().await;
-        let app_config = create_test_app_config();
-        let app = create_app_with_config(app_state, app_config);
-
-        let request = Request::builder()
-            .method(Method::GET)
-            .uri("/api/v1/admin/system")
-            .body(Body::empty())
-            .unwrap();
-
-        let response = app.oneshot(request).await.unwrap();
-        assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn test_websocket_route_without_auth() {
-        let app_state = create_test_app_state().await;
-        let app_config = create_test_app_config();
-        let app = create_app_with_config(app_state, app_config);
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method(Method::GET)
-                    .uri("/ws")
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(response.status().as_u16(), 404);
-    }
-
-    #[tokio::test]
-    async fn test_route_structure() {
-        let app_state = create_test_app_state().await;
-        let app_config = create_test_app_config();
-        let app = create_app_with_config(app_state, app_config);
-
-        let health_request = Request::builder()
-            .method(Method::GET)
-            .uri("/health")
-            .body(Body::empty())
-            .unwrap();
-
-        let response = app.oneshot(health_request).await.unwrap();
-        assert_eq!(response.status(), StatusCode::OK);
-    }
-
-    #[test]
-    fn test_health_check_response() {
-        let _response = health_check();
-        // Health check compilation test - compiles successfully
-    }
-
-    #[tokio::test]
-    async fn test_app_state_cloning() {
-        let app_state = create_test_app_state().await;
-
-        let backend_router_clone = app_state.backend_router.clone();
-        let jwt_service_clone = app_state.jwt_service.clone();
-        let user_store_clone = app_state.user_store.clone();
-
-        assert!(Arc::ptr_eq(
-            &app_state.backend_router,
-            &backend_router_clone
-        ));
-        assert!(Arc::ptr_eq(&app_state.jwt_service, &jwt_service_clone));
-        assert!(Arc::ptr_eq(&app_state.user_store, &user_store_clone));
-    }
-
-    #[tokio::test]
-    async fn test_app_state_structure() {
-        let app_state = create_test_app_state().await;
-
-        assert!(Arc::strong_count(&app_state.backend_router) >= 1);
-        assert!(Arc::strong_count(&app_state.jwt_service) >= 1);
-        assert!(Arc::strong_count(&app_state.user_store) >= 1);
-    }
-
-    #[tokio::test]
-    async fn test_json_rejection_handling() {
-        let app_state = create_test_app_state().await;
-        let app_config = create_test_app_config();
-        let app = create_app_with_config(app_state, app_config);
-
-        // Test invalid JSON handling on existing auth route
-        let request = Request::builder()
-            .method(Method::POST)
-            .uri("/auth/login")
-            .header("content-type", "application/json")
-            .body(Body::from("invalid json"))
-            .unwrap();
-
-        let response = app.oneshot(request).await.unwrap();
-        assert_eq!(response.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn test_api_docs_endpoint() {
-        let app_state = create_test_app_state().await;
-        let app_config = create_test_app_config();
-        let app = create_app_with_config(app_state, app_config);
-
-        let request = Request::builder()
-            .method(Method::GET)
-            .uri("/docs")
-            .body(Body::empty())
-            .unwrap();
-
-        let response = app.oneshot(request).await.unwrap();
-        // API docs might not be implemented, so we just check it doesn't crash
-        assert!(response.status().is_client_error() || response.status().is_success());
-    }
-
-    #[tokio::test]
-    async fn test_app_state_error_handling() {
-        let app_state = create_test_app_state().await;
-        let app_config = create_test_app_config();
-        let app = create_app_with_config(app_state, app_config);
-
-        // Test error handling with malformed request
-        let request = Request::builder()
-            .method(Method::POST)
-            .uri("/api/v1/data/test")
-            .header("content-type", "application/json")
-            .body(Body::from("malformed"))
-            .unwrap();
-
-        let response = app.oneshot(request).await.unwrap();
-        // Should return either 401 (unauthorized) or 400 (bad request)
-        assert!(
-            response.status() == StatusCode::UNAUTHORIZED
-                || response.status() == StatusCode::BAD_REQUEST
-        );
-    }
-}
